Split StatusWord in a separate file
