+++ tests/ragger/test_eip7702.py
@@ -4,7 +4,8 @@ from ragger.error import ExceptionRAPDU
 from ragger.navigator.navigation_scenario import NavigateWithScenario
 from ragger.navigator import NavInsID
 
-from client.client import EthAppClient, StatusWord
+from client.client import EthAppClient
+from client.status_word import StatusWord
 from client.settings import SettingID, settings_toggle
 import client.response_parser as ResponseParser
 from client.tx_auth_7702 import TxAuth7702
